+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## _v1.8.0_
+
+### **Date: 23-MAY-2023**
+
+- AZURE_EU, Region support added
+- Include Metadata support added to asset, entry and query
+- General code improvement clean up
+- Updated code for Live Preview
+
 ---
 
 ## _v1.7.0_
@@ -67,13 +76,13 @@ Date: 08-Dec-2020
 - Timeout support included
 
 - Entry
-  - added support for include_fallback.
+    - added support for include_fallback.
 - Asset
-  - added support for include_fallback.
+    - added support for include_fallback.
 - AssetQueryable
-  - added support for include_fallback.
+    - added support for include_fallback.
 - Query
-  - added support for include_fallback.
+    - added support for include_fallback.
 
 ---