+++ changelog.rst
@@ -2,11 +2,21 @@
 **CHANGELOG**
 ================
 
-*v1.7.0*
-
+*v1.8.0*
 ============
 
-**Date: 8-APR-2022**
+**Date: 23-MAY-2023**
+
+- AZURE_EU, Region support added
+- Include Metadata support added to asset, entry and query
+- General code improvement clean up
+- Updated code for Live Preview
+
+
+*v1.7.0*
+============
+
+**Date: 08-APR-2022**
 
 Region support added.
     - AZURE_NA support added